+++ avr_controlRemoto/src/library/tranformVar/transformVar.hpp
@@ -3,8 +3,8 @@
 #include <avr/io.h>
 #include <util/delay.h>
 int char_to_int(char *);
-int numeroUnidades(long );
-void int_to_char(long ,char *,int );
+int numeroUnidades(int );
+void int_to_char(int ,char *,int );
 int int_to_binario(int8_t  numero);
 
 #endif
